requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.1
attrs==25.4.0
colorzero==2.0
frozenlist==1.8.0
gpiozero==2.0.1
idna==3.11
lgpio==0.2.2.0
multidict==6.7.0
numpy==2.3.4
packaging==25.0
phonenumbers==9.0.18
pillow==12.0.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
rpi-lgpio==0.6
setuptools==80.9.0
signalbot==0.20.0
spidev==3.8
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
websockets==15.0.1
yarl==1.22.0
